Convert more classes to use Lombok
- Also move dtos into a package
