+++ src/main/java/application/models/dto/QuestionDTO.java
@@ -1,13 +1,17 @@
-package application.models;
+package application.models.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.*;
 
 import java.util.Collection;
-import java.util.Objects;
 
 /**
  * A Data Transfer Object representing a Question Object
  */
+@Getter
+@Setter
+@NoArgsConstructor
+@EqualsAndHashCode
 public class QuestionDTO {
 
     private String questionType;
@@ -23,8 +27,6 @@ public class QuestionDTO {
 
     public static final String OPENENDED = "openEnded", RANGE = "range", MULTIPLECHOICE = "multipleChoice";
 
-    public QuestionDTO(){}
-
     public QuestionDTO(String questionType, String question, int min, int max, Collection<String> choices, Long id) {
         this.questionType = questionType;
         this.question = question;
@@ -33,33 +35,6 @@ public class QuestionDTO {
         this.choices = choices;
         this.ID =id;
     }
-
-    public void setID(Long ID) {
-        this.ID = ID;
-    }
-    public Long getID() {
-        return ID;
-    }
-
-    public String getQuestionType() { return questionType; }
-
-    public void setQuestionType(String questionType) { this.questionType = questionType; }
-
-    public String getQuestion() { return question; }
-
-    public void setQuestion(String question) { this.question = question; }
-
-    public int getMin() { return min; }
-
-    public void setMin(int min) { this.min = min; }
-
-    public int getMax() { return max; }
-
-    public void setMax(int max) { this.max = max; }
-
-    public Collection<String> getChoices() { return choices; }
-
-    public void setChoices(Collection<String> choices) { this.choices = choices; }
 
     @Override
     public String toString(){
@@ -79,17 +54,5 @@ public class QuestionDTO {
              break;
         }
     return str;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        QuestionDTO that = (QuestionDTO) o;
-        return min == that.min &&
-                max == that.max &&
-                Objects.equals(questionType, that.questionType) &&
-                Objects.equals(question, that.question) &&
-                Objects.equals(choices, that.choices);
     }
 }
