+++ src/test/java/application/controllers/TestSurveyController.java
@@ -1,6 +1,9 @@
 package application.controllers;
 
 import application.models.*;
+import application.models.dto.AnswerDTO;
+import application.models.dto.QuestionDTO;
+import application.models.dto.SurveyDTO;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.*;
 import org.mockito.InjectMocks;
