+++ src/test/java/application/models/TestAnswerDTO.java
@@ -1,5 +1,6 @@
 package application.models;
 
+import application.models.dto.AnswerDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
