+++ src/test/java/application/models/TestEditDTO.java
@@ -1,12 +1,11 @@
 package application.models;
 
+import application.models.dto.EditDTO;
+import application.models.dto.QuestionDTO;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
+import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -44,7 +43,7 @@ public class TestEditDTO {
 
     @Test
     public void constructorTest(){
-        editDTO = new EditDTO(TEST_SURVEY_ID,TEST_NEW_QUESTION_DTO_LIST, TEST_EDITED_QUESTIONS, TEST_SURVEY_NAME);
+        editDTO = new EditDTO(TEST_SURVEY_ID, TEST_SURVEY_NAME, TEST_NEW_QUESTION_DTO_LIST, TEST_EDITED_QUESTIONS);
         assertEquals(TEST_SURVEY_NAME, editDTO.getSurveyName());
         assertEquals(10, editDTO.getSurveyID());
         assertEquals(TEST_NEW_QUESTION_DTO_LIST, editDTO.getNewQuestions());
